Rename to createWishlistForm
