+++ packages/venia-ui/i18n/en_US.json
@@ -221,10 +221,10 @@
     "navHeader.myAccountText": "My Account",
     "navHeader.signInText": "Sign In",
     "navigationTrigger.ariaLabel": "Toggle navigation panel",
-    "newWishlistForm.privateRadio": "Private",
-    "newWishlistForm.publicRadio": "Public",
-    "newWishlistForm.cancelButton": "Cancel",
-    "newWishlistForm.saveButton": "Save",
+    "createWishlistForm.privateRadio": "Private",
+    "createWishlistForm.publicRadio": "Public",
+    "createWishlistForm.cancelButton": "Cancel",
+    "createWishlistForm.saveButton": "Save",
     "noProductsFound.noProductsFound": "Sorry! We couldn't find any products.",
     "noProductsFound.tryOneOfTheseCategories": "Try one of these categories",
     "Order Status": "Order Status",
