+++ packages/venia-ui/lib/components/Wishlist/WishlistDialog/CreateWishlistForm/index.js
@@ -1,0 +1,1 @@
+export { default } from './createWishlistForm';
