--- packages/venia-ui/lib/components/Wishlist/WishlistDialog/NewWishlistForm/index.js
@@ -1,1 +1,0 @@
-export { default } from './newWishlistForm';
